+++ packages/root/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+- Rename `raw` option to `pathname`.
+
 ## 2.0.0-beta.2
 
 - Remove `location.rawPathname`. Rely on the `raw` option to correctly format the provided pathname.